feat(bh_status): unround Qty, add FillAvg column + executions cache
bh_status was rounding fractional share quantities to int and reporting
IBKR's commission-inclusive avgCost — making the table impossible to
reconcile against the post-fill emails. Now Qty renders fractional
shares verbatim and a new FillAvg column shows the weighted-avg raw
fill price (commissions excluded) from a persistent local cache at
src/logs/ibkr_executions_cache.json. The cache is needed because IBKR's
reqExecutions only returns the current session regardless of the time
filter, so older fills would otherwise show "—" forever; every run now
merges that day's fills in by exec_id and reuses them on subsequent
runs.

Also fixes get_executions casting execution.shares to int, which was
silently truncating fractional fills.

+++ src/bluehorseshoe/data/ibkr_client.py
@@ -324,7 +324,7 @@ class IBKRClient:
                     "exec_id": execution.execId,
                     "symbol": fill.contract.symbol,
                     "side": execution.side.lower(),  # "BOT" → "bot", "SLD" → "sld"
-                    "quantity": int(execution.shares),
+                    "quantity": float(execution.shares),
                     "price": float(execution.price),
                     "commission": commission,
                     "exec_time": execution.time,
